requirements-gpu.txt
tensorflow-gpu==2.3.0rc0
opencv-python==4.1.1.26
lxml
tqdm
absl-py
matplotlib
easydict
pillow
pytesseract
keras
sklearn
flask
flask-cors